Made it nearter and shorter responses
