+++ src/main/java/org/ieeervce/gatekeeper/controller/SocietyController.java
@@ -1,7 +1,7 @@
 package org.ieeervce.gatekeeper.controller;
 
 import org.ieeervce.gatekeeper.exception.ItemNotFoundException;
-import org.ieeervce.gatekeeper.dto.SocietyDTO;
+import org.ieeervce.gatekeeper.dto.Society.SocietyDTO;
 import org.ieeervce.gatekeeper.entity.Society;
 import org.ieeervce.gatekeeper.service.SocietyService;
 import org.modelmapper.ModelMapper;
