+++ src/main/java/org/ieeervce/gatekeeper/dto/RequestForm/RequestFormPdfDTO.java
@@ -1,4 +1,4 @@
-package org.ieeervce.gatekeeper.dto;
+package org.ieeervce.gatekeeper.dto.RequestForm;
 
 public class RequestFormPdfDTO {
     private byte[] formPDF;
