+++ src/main/java/org/ieeervce/gatekeeper/dto/RequestForm/ResponseRequestFormDTO.java
@@ -1,14 +1,9 @@
-package org.ieeervce.gatekeeper.dto;
+package org.ieeervce.gatekeeper.dto.RequestForm;
 
-import jakarta.persistence.*;
-import org.hibernate.annotations.ColumnDefault;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
+import org.ieeervce.gatekeeper.dto.ReviewLog.ReviewLogDTO;
+import org.ieeervce.gatekeeper.dto.User.UserInternalDTO;
 import org.ieeervce.gatekeeper.entity.FinalStatus;
-import org.ieeervce.gatekeeper.entity.ReviewLog;
 import org.ieeervce.gatekeeper.entity.Role;
-import org.ieeervce.gatekeeper.entity.User;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -21,7 +16,7 @@ public class ResponseRequestFormDTO {
     private String eventTitle;
     private int formValue;
 
-    private User requester;
+    private UserInternalDTO requester;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 
@@ -75,11 +70,11 @@ public class ResponseRequestFormDTO {
         this.eventTitle = eventTitle;
     }
 
-    public User getRequester() {
+    public UserInternalDTO getRequester() {
         return requester;
     }
 
-    public void setRequester(User requester) {
+    public void setRequester(UserInternalDTO requester) {
         this.requester = requester;
     }
