--- src/main/java/org/ieeervce/gatekeeper/dto/SocietyDTO.java
@@ -1,17 +1,0 @@
-package org.ieeervce.gatekeeper.dto;
-
-/**
- * Society DTO to create/update a society.
- */
-public class SocietyDTO {
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-}
